requirements.txt
# FloraWatch Backend - Dependencias Básicas
# Instalar primero estas dependencias básicas

# Django y REST Framework
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Utilidades básicas
python-dotenv>=1.0.0
requests>=2.28.0
pillow>=10.0.0

# Base de datos PostgreSQL + PostGIS
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# GIS Support
GDAL>=3.7.0
Fiona>=1.9.0
Shapely>=2.0.0
pyproj>=3.6.0
djangorestframework-gis>=1.0

# Machine Learning básico
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Desarrollo y testing
pytest>=7.0.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=7.3.0
